data/crawlers/naver_place.py
"""네이버 플레이스 음식점 크롤러"""

import requests
from bs4 import BeautifulSoup
from dataclasses import dataclass


@dataclass
class Restaurant:
    name: str
    address: str
    lat: float
    lng: float
    category: str
    phone: str
    menu: list[dict]
    rating: float
    review_count: int
    source: str = "naver_place"


class NaverPlaceCrawler:
    BASE_URL = "https://map.naver.com/v5/api/search"

    def __init__(self):
        self.session = requests.Session()
        self.session.headers.update({
            "User-Agent": "Mozilla/5.0",
            "Referer": "https://map.naver.com/",
        })

    def search(self, query: str, region: str = "강남구") -> list[Restaurant]:
        """음식점 검색"""
        # TODO: 실제 크롤링 로직 구현
        print(f"Searching: {query} in {region}")
        return []

    def get_detail(self, place_id: str) -> Restaurant | None:
        """음식점 상세 정보"""
        # TODO: 상세 페이지 크롤링
        return None


if __name__ == "__main__":
    crawler = NaverPlaceCrawler()
    results = crawler.search("맛집")
    print(f"Found {len(results)} restaurants")
